refactor(constant): Refactor Platform File
