+++ lib/constants/platforms.ts
